refactor of form into two components
